refactor: Layout 컴포넌트 common 폴더로 이동 및 적용
