+++ pages/index.tsx
@@ -15,8 +15,8 @@ import useSWRInfinite from 'swr/infinite';
 import useMutation from '@libs/client/useMutation';
 import { ProfileResponse } from '@libs/client/useUser';
 import client from '@libs/server/client';
+import Layout from '@components/common/Layout';
 import HeadMeta from '@components/headMeta';
-import Layout from '@components/layout';
 import LoadingSpinner from '@components/loadingSpinner';
 import DeleteModal from '@components/profile/deleteModal';
 import ClickedProject from '@components/project/clickedProject';
