+++ pages/joblist/index.tsx
@@ -2,7 +2,7 @@ import type { NextPage } from 'next';
 import { useRouter } from 'next/router';
 import useSWR from 'swr';
 import { useUserState } from '@libs/client/useUser';
-import Layout from '@components/layout';
+import Layout from '@components/common/Layout';
 
 const Joblist: NextPage = () => {
   const { data, error }: useUserState = useSWR('/api/users/me');
@@ -11,7 +11,8 @@ const Joblist: NextPage = () => {
     <Layout
       isLogin={data && data.ok}
       profile={data?.profile}
-      userId={data?.profile?.id}>
+      userId={data?.profile?.id}
+    >
       <div>안녕</div>
     </Layout>
   );
