+++ pages/live/[id]/index.tsx
@@ -2,7 +2,7 @@ import type { NextPage } from 'next';
 import { useRouter } from 'next/router';
 import useSWR from 'swr';
 import { useUserState } from '@libs/client/useUser';
-import Layout from '@components/layout';
+import Layout from '@components/common/Layout';
 import InteractionButton from '@components/stream/interactionButton';
 
 const Live: NextPage = () => {
@@ -12,7 +12,8 @@ const Live: NextPage = () => {
     <Layout
       isLogin={data && data.ok}
       profile={data?.profile}
-      userId={data?.profile?.id}>
+      userId={data?.profile?.id}
+    >
       <div className="relative flex">
         <div className="w-full overflow-scroll">
           <div className="aspect-video bg-slate-200"></div>
@@ -28,7 +29,8 @@ const Live: NextPage = () => {
                     xmlns="http://www.w3.org/2000/svg"
                     className="h-4 w-4"
                     viewBox="0 0 20 20"
-                    fill="currentColor">
+                    fill="currentColor"
+                  >
                     <path d="M10 12a2 2 0 100-4 2 2 0 000 4z" />
                     <path
                       fillRule="evenodd"
