+++ pages/live/index.tsx
@@ -3,7 +3,7 @@ import { useRouter } from 'next/router';
 import useSWR from 'swr';
 import { ProjectsResponse } from 'pages';
 import { useUserState } from '@libs/client/useUser';
-import Layout from '@components/layout';
+import Layout from '@components/common/Layout';
 
 const Live: NextPage = () => {
   const { data, error }: useUserState = useSWR('/api/users/me');
@@ -12,7 +12,8 @@ const Live: NextPage = () => {
     <Layout
       isLogin={data && data.ok}
       profile={data?.profile}
-      userId={data?.profile?.id}>
+      userId={data?.profile?.id}
+    >
       <div className="grid grid-cols-5 gap-6 px-6 py-6">
         <div className="cursor-pointer">
           <div className="rounded-md bg-gradient-to-t from-black/100 to-black/50">
