+++ pages/portfolio/index.tsx
@@ -1,7 +1,7 @@
 import type { NextPage } from 'next';
 import useSWR from 'swr';
 import { useUserState } from '@libs/client/useUser';
-import Layout from '@components/layout';
+import Layout from '@components/common/Layout';
 
 const Live: NextPage = () => {
   const { data }: useUserState = useSWR('/api/users/me');
@@ -10,7 +10,8 @@ const Live: NextPage = () => {
     <Layout
       isLogin={data && data.ok}
       profile={data?.profile}
-      userId={data?.profile?.id}>
+      userId={data?.profile?.id}
+    >
       <div>안녕</div>
     </Layout>
   );
