+++ pages/profile/[id]/editor.tsx
@@ -9,8 +9,8 @@ import { makeImageURL } from '@libs/client/utils';
 import Button from '@components/common/Button';
 import ErrorMessage from '@components/common/ErrorMessage';
 import Input from '@components/common/Input';
+import Layout from '@components/common/Layout';
 import HeadMeta from '@components/headMeta';
-import Layout from '@components/layout';
 import CategoryTabMenu from '@components/profile/categoryTabMenu';
 import ProfileWeb from '@components/profile/profileWeb';
 import TextArea from '@components/textArea';
