+++ pages/user/create.tsx
@@ -14,8 +14,8 @@ import Button from '@components/common/Button';
 import ErrorMessage from '@components/common/ErrorMessage';
 import Input from '@components/common/Input';
 import InputPassword from '@components/common/InputPassword';
+import Layout from '@components/common/Layout';
 import HeadMeta from '@components/headMeta';
-import Layout from '@components/layout';
 import APPLE_LOGO from '@public/apple.png';
 import FACEBOOK_LOGO from '@public/facebook.svg';
 import GOOGLE_LOGO from '@public/google.svg';
